[fix]元のtopページを削除しルートに移植、footerとheaderのtopリンクを変更
