+++ src/app/page.tsx
@@ -1,101 +1,63 @@
-import Image from "next/image";
+import BackFrame from '@/src/components/common/back_frame';
+import Line from '@/src/components/common/line';
+import TextStyle from '@/src/components/common/text_style';
+import Infos from '@/src/components/top/Infos';
+import RecruitmentSponsors from '@/src/components/top/recruitment_sponsors';
+import TopAttentions from '@/src/components/top/top_attentions';
+import TopCarousel from '@/src/components/top/top_carousel';
+import Image from 'next/image';
+export const runtime = 'edge';
 
-export default function Home() {
+export default function TopPage() {
   return (
-    <div className="grid grid-rows-[20px_1fr_20px] items-center justify-items-center min-h-screen p-8 pb-20 gap-16 sm:p-20 font-[family-name:var(--font-geist-sans)]">
-      <main className="flex flex-col gap-8 row-start-2 items-center sm:items-start">
-        <Image
-          className="dark:invert"
-          src="https://nextjs.org/icons/next.svg"
-          alt="Next.js logo"
-          width={180}
-          height={38}
-          priority
-        />
-        <ol className="list-inside list-decimal text-sm text-center sm:text-left font-[family-name:var(--font-geist-mono)]">
-          <li className="mb-2">
-            Get started by editing{" "}
-            <code className="bg-black/[.05] dark:bg-white/[.06] px-1 py-0.5 rounded font-semibold">
-              app/page.tsx
-            </code>
-            .
-          </li>
-          <li>Save and see your changes instantly.</li>
-        </ol>
-
-        <div className="flex gap-4 items-center flex-col sm:flex-row">
-          <a
-            className="rounded-full border border-solid border-transparent transition-colors flex items-center justify-center bg-foreground text-background gap-2 hover:bg-[#383838] dark:hover:bg-[#ccc] text-sm sm:text-base h-10 sm:h-12 px-4 sm:px-5"
-            href="https://vercel.com/new?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <Image
-              className="dark:invert"
-              src="https://nextjs.org/icons/vercel.svg"
-              alt="Vercel logomark"
-              width={20}
-              height={20}
-            />
-            Deploy now
-          </a>
-          <a
-            className="rounded-full border border-solid border-black/[.08] dark:border-white/[.145] transition-colors flex items-center justify-center hover:bg-[#f2f2f2] dark:hover:bg-[#1a1a1a] hover:border-transparent text-sm sm:text-base h-10 sm:h-12 px-4 sm:px-5 sm:min-w-44"
-            href="https://nextjs.org/docs?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            Read our docs
-          </a>
+    <div>
+      <div style={{ paddingBottom: '308px' }} className="space-y-6 pt-4">
+        <div className="flex justify-center">
+          <Image
+            src="/logo/title_logo.svg"
+            alt="44th_logo"
+            width={200}
+            height={200}
+            className="slide-in-bck-top"
+          />
         </div>
-      </main>
-      <footer className="row-start-3 flex gap-6 flex-wrap items-center justify-center">
-        <a
-          className="flex items-center gap-2 hover:underline hover:underline-offset-4"
-          href="https://nextjs.org/learn?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
+        <div className="flex justify-center">
           <Image
-            aria-hidden
-            src="https://nextjs.org/icons/file.svg"
-            alt="File icon"
-            width={16}
-            height={16}
+            src="/logo/title_date.svg"
+            alt="44th_date"
+            width={200}
+            height={200}
+            className="slide-in-bck-top"
           />
-          Learn
-        </a>
-        <a
-          className="flex items-center gap-2 hover:underline hover:underline-offset-4"
-          href="https://vercel.com/templates?framework=next.js&utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <Image
-            aria-hidden
-            src="https://nextjs.org/icons/window.svg"
-            alt="Window icon"
-            width={16}
-            height={16}
-          />
-          Examples
-        </a>
-        <a
-          className="flex items-center gap-2 hover:underline hover:underline-offset-4"
-          href="https://nextjs.org?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <Image
-            aria-hidden
-            src="https://nextjs.org/icons/globe.svg"
-            alt="Globe icon"
-            width={16}
-            height={16}
-          />
-          Go to nextjs.org →
-        </a>
-      </footer>
+        </div>
+      </div>
+      <BackFrame>
+        <div className="py-4 space-y-8">
+          <TopCarousel />
+          <RecruitmentSponsors />
+          <Infos />
+          <TopAttentions />
+          <Line />
+
+          <div className="text-center">
+            <TextStyle styleType="section_title">タイムスケジュール</TextStyle>
+          </div>
+
+          <div className="text-center">
+            <TextStyle styleType="body2">Coming soon</TextStyle>
+          </div>
+
+          <Line />
+          <div className="text-center">
+            {' '}
+            <TextStyle styleType="section_title">関連アプリ・サイト</TextStyle>
+          </div>
+
+          <div className="text-center">
+            <TextStyle styleType="body2">Coming soon</TextStyle>
+          </div>
+        </div>
+      </BackFrame>
     </div>
   );
 }
